fix(render): the derived path reserve mixed usize columns into u16 arithmetic
252735a0 replaced the hardcoded `saturating_sub(6)` with `saturating_sub(path_label.width() + 1)`,
but `UnicodeWidthStr::width` returns `usize` and `paint_path_line`'s `width` is a `u16`, so
`u16::saturating_sub` had no matching impl. The cast was already there for the result
(`as usize`); it was missing on the argument.

Checked the other three `width()` sites in the same change for the same mistake:
`meta.width() + title_text.width() + 6 < overlay_rect.width as usize`, and the two
`... as u16` casts in the overlay title and loading label are usize-in/u16-out and
keep their original types.

+++ crates/codegen/xai-grok-pager-render/src/render/image_overlay/content.rs
@@ -27,7 +27,10 @@ pub(super) fn paint_path_line(
     let label = format!(
         "{} {}",
         path_label,
-        truncate_path_for_overlay(&raw, width.saturating_sub(path_label.width() + 1) as usize)
+        truncate_path_for_overlay(
+            &raw,
+            width.saturating_sub((path_label.width() + 1) as u16) as usize,
+        )
     );
     let clipped = crate::render::line_utils::truncate_str(&label, width as usize);
     buf.set_span_safe(
